+++ cli/eproto/eproto.cc
@@ -50,7 +50,8 @@ int main(int argc, char* argv[]) {
   argparse::ArgumentParser program("vlink-eproto", VLINK_VERSION, argparse::default_arguments::all);
 
   program.add_description("Note: You may need to add multicast/broadcast [" +
-                          vlink::DiscoveryViewer::get_listen_address() + "]");
+                          vlink::DiscoveryViewer::get_listen_address() + ":" +
+                          std::to_string(vlink::DiscoveryViewer::get_listen_port()) + "]");
 
   argparse::ArgumentParser pub_command("pub", VLINK_VERSION, argparse::default_arguments::help);
   pub_command.add_argument("url").help("Bind url").required();