+++ cli/list/list.cc
@@ -362,7 +362,8 @@ int main(int argc, char* argv[]) {
   argparse::ArgumentParser program("vlink-list", VLINK_VERSION, argparse::default_arguments::all);
 
   program.add_description("Note: You may need to add multicast/broadcast [" +
-                          vlink::DiscoveryViewer::get_listen_address() + "]");
+                          vlink::DiscoveryViewer::get_listen_address() + ":" +
+                          std::to_string(vlink::DiscoveryViewer::get_listen_port()) + "]");
 
   program.add_argument("-n", "--native").help("Native mode").default_value(false).implicit_value(true);
   program.add_argument("-m", "--name").help("Process name").default_value(std::string());